# docs/swagger.yaml
basePath: /
definitions:
  model.Currency:
    properties:
      code:
        type: string
      date:
        type: string
      id:
        type: integer
      title:
        type: string
      value:
        type: number
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
host: localhost:8000
info:
  contact: {}
  title: CurrencyService
  version: "1.0"
paths:
  /currency/{date}:
    get:
      consumes:
      - application/json
      description: Get currency data by date and code
      parameters:
      - description: Date in format YYYY.MM.DD
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Currency'
            type: array
      summary: Get currency data
      tags:
      - currency
  /currency/{date}/{code}:
    get:
      consumes:
      - application/json
      description: Get currency data by date and code
      parameters:
      - description: Date in format YYYY.MM.DD
        in: path
        name: date
        required: true
        type: string
      - description: Currency code (optional)
        in: path
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Currency'
            type: array
      summary: Get currency data
      tags:
      - currency
  /currency/save/{date}:
    get:
      consumes:
      - application/json
      description: Save currency data by date
      parameters:
      - description: Date in format YYYY.MM.DD
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Save
      tags:
      - currency
swagger: "2.0"